+++ tests/CMakeLists.txt
@@ -13,6 +13,7 @@ include_directories(
 # First thing define the common source:
 set(compare_images_SRCS compare_images.c
   ${OPENJPEG_SOURCE_DIR}/src/bin/jp2/convert.c
+  ${OPENJPEG_SOURCE_DIR}/src/bin/jp2/converttif.c
   ${OPENJPEG_SOURCE_DIR}/src/bin/common/opj_getopt.c
   )
